UC-1223 adding eddsa support
