+++ examples/verification.verifyFromKeyService.js
@@ -2,9 +2,9 @@ const verify = require("../src/verify-with-key-service");
 // Use this if installed from NPM
 // const verify = require("@ubirch/ubirch-protocol-verifier");
 
-const stage = "prod";
-const uuid = '7af55a89-d9a0-4929-b5aa-1b1657b06a56';
-const upp = "liPEEHr1WonZoEkptaobFlewalbEQCI+PfqAymyGmdyszoDVDvQwTTs9aSLfDwFr163jyiXpKfpaddpkR6g7DtfaCz/4IJyRLostPO2PWsgwigqW9G8AxCB01pV0Aw5ngFA9RXBpJ8nu+dT8chghOj1goS4O38ZWisRATv0reU41YtXKJp6lpXh5Jt5buq4n17sBbVm3GLyiAeTmSGuAEQcxbE7j7UhOQLg8uR1Oj/Ql2tbKmhRCzR5jWg==";
+const stage = "dev";
+const uuid = 'c3e59fee-c409-4ab7-9aae-b19e62f4af9b';
+const upp = "lSLEEMPln+7ECUq3mq6xnmL0r5sAxCCzns3jSfHQrt6sQA62/GD6Dv2cYyINCxzAU/cT3UOZa8RAMkNE2PGYIw63/lmbLNpAE/OIAO2WrjNQXqt9mszz0IcdGwZHocCyCCzj7LpP9WcRPIN1LdvS/tsLBWWgEsXFAA==";
 
 verify.verifyFromKeyService(stage, uuid, upp).then((res) => {
     console.log(res);
