+++ package.json
@@ -31,6 +31,11 @@
   "author": "Carlos Sanchez",
   "license": "Apache-2.0",
   "devDependencies": {
+    "eslint": "^8.18.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.2.3",
+    "eslint-plugin-promise": "^6.0.0",
     "typescript": "^4.2.4",
     "webpack": "^5.17.0",
     "webpack-cli": "^4.4.0"
